inherit the default behaviour of spring for internal object mapper
